tests/requirements.txt
pytest
selene>=2.0.b4
allure-pytest
